+++ elfo-core/src/object.rs
@@ -29,12 +29,6 @@ pub(crate) enum ObjectKind {
     Group(Group),
 }
 
-#[derive(Debug, Hash, PartialEq)]
-pub struct ObjectMeta {
-    pub group: String,
-    pub key: Option<String>,
-}
-
 impl Object {
     pub(crate) fn new(addr: Addr, kind: impl Into<ObjectKind>) -> Self {
         Self {