+++ elfo-logger/src/lib.rs
@@ -14,7 +14,7 @@ use sharded_slab::Pool;
 use tracing::{span::Id as SpanId, Level, Metadata, Subscriber};
 use tracing_subscriber::{prelude::*, registry::Registry, EnvFilter};
 
-use elfo_core::{trace_id::TraceId, Schema, _priv::ObjectMeta};
+use elfo_core::{trace_id::TraceId, ActorMeta, Schema};
 
 use crate::{actor::Logger, filtering_layer::FilteringLayer, printing_layer::PrintingLayer};
 
@@ -48,7 +48,7 @@ struct PreparedEvent {
     timestamp: SystemTime,
     trace_id: Option<TraceId>,
     metadata: &'static Metadata<'static>,
-    object: Option<Arc<ObjectMeta>>,
+    object: Option<Arc<ActorMeta>>,
     span_id: Option<SpanId>,
     payload_id: StringId,
 }