+++ elfo-logger/src/theme.rs
@@ -2,7 +2,7 @@ use std::{sync::Arc, time::SystemTime};
 
 use tracing::Level;
 
-use elfo_core::{_priv::ObjectMeta, trace_id::TraceId};
+use elfo_core::{trace_id::TraceId, ActorMeta};
 
 use crate::formatters::*;
 
@@ -10,7 +10,7 @@ pub(crate) trait Theme {
     type Timestamp: Formatter<SystemTime>;
     type Level: Formatter<Level>;
     type TraceId: Formatter<Option<TraceId>>;
-    type ObjectMeta: Formatter<Option<Arc<ObjectMeta>>>;
+    type ActorMeta: Formatter<Option<Arc<ActorMeta>>>;
     type Payload: Formatter<str>;
     type Location: Formatter<(&'static str, u32)>;
     type Module: Formatter<str>;
@@ -19,10 +19,10 @@ pub(crate) trait Theme {
 pub(crate) struct PlainTheme;
 
 impl Theme for PlainTheme {
+    type ActorMeta = EmptyIfNone<Arc<ActorMeta>>;
     type Level = Level;
     type Location = Location;
     type Module = Module;
-    type ObjectMeta = EmptyIfNone<Arc<ObjectMeta>>;
     type Payload = Payload;
     type Timestamp = Rfc3339Weak;
     type TraceId = EmptyIfNone<TraceId>;
@@ -31,10 +31,10 @@ impl Theme for PlainTheme {
 pub(crate) struct ColoredTheme;
 
 impl Theme for ColoredTheme {
+    type ActorMeta = EmptyIfNone<ColoredByHash<Arc<ActorMeta>>>;
     type Level = ColoredLevel;
     type Location = ColoredLocation;
     type Module = ColoredModule;
-    type ObjectMeta = EmptyIfNone<ColoredByHash<Arc<ObjectMeta>>>;
     type Payload = ColoredPayload;
     type Timestamp = Rfc3339Weak;
     type TraceId = EmptyIfNone<ColoredByHash<TraceId>>;