+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Expose `ActorMeta` and `ActorStatusKind`.
+
 ### Changed
 - context: `Context::recv()` and `Context::try_recv()` panics if called again after returning `None`.
 - logger: use `_location` and `_module` instead of `@location` and `@module`.